test adjusted to the code update
